rename visualization.py to plotting.py
