+++ kenchi/__init__.py
@@ -2,7 +2,7 @@ from . import datasets
 from . import metrics
 from . import outlier_detection
 from . import pipeline
-from . import visualization
+from . import plotting
 
 __version__ = '0.9.0'
 
@@ -11,6 +11,6 @@ __all__     = [
     'metrics',
     'outlier_detection',
     'pipeline',
-    'visualization',
+    'plotting',
     '__version__'
 ]
